Added documentation to the Server class. Removed unused variables.
