+++ src/main/kotlin/network/Server.kt
@@ -6,6 +6,8 @@ import data.communication.TransmissionType
 import data.network.Endpoint
 import data.network.Node
 import kademlia.Kademlia
+import kotlinx.serialization.encodeToByteArray
+import kotlinx.serialization.protobuf.ProtoBuf
 import logging.Dashboard
 import logging.Logger
 import utils.Utils.Companion.asHex
@@ -33,15 +35,16 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
 
     val isTrustedNode = localNode.let { node -> node.ip == configuration.trustedNodeIP && node.port == configuration.trustedNodePort }
 
-    private val receivedQueue = LinkedBlockingQueue<MessageBuilder>()
+    private val processingQueue = LinkedBlockingQueue<MessageBuilder>()
     private val outgoingQueue = LinkedBlockingQueue<OutgoingQueuedMessage>()
-    private val queuedForLater = ConcurrentHashMap<String, OutgoingQueuedMessage>()
 
-    private val networkHistory = ConcurrentHashMap<String, Long>()
+    private val messageHistory = ConcurrentHashMap<String, Long>()
     private val messageBuilders = mutableMapOf<String, MessageBuilder>()
     private val udpSocket = DatagramSocket(configuration.port)
     protected val tcpSocket = ServerSocket(configuration.port + 1)
     private var started = false
+
+    abstract fun onMessageReceived(endpoint: Endpoint, data: ByteArray)
 
     open fun launch() {
         if (isTrustedNode) Logger.info("We're the trusted node.")
@@ -53,6 +56,7 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
         started = true
     }
 
+    /**Listens for UDP [messages][Message] and adds them to the processing queue. If the Message is [TransmissionType.Broadcast] it is broadcasted. */
     private fun listenForUDP() {
         val pureArray = ByteArray(configuration.packetSplitSize)
         val inputStream = DataInputStream(ByteArrayInputStream(pureArray))
@@ -62,8 +66,8 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
             udpSocket.receive(packet)
             val packetId = inputStream.readNBytes(32).asHex
             val messageId = inputStream.readNBytes(32).asHex
-            if (networkHistory.containsKey(packetId) || networkHistory.containsKey(messageId)) return@tryAndReport
-            networkHistory[packetId] = System.currentTimeMillis()
+            if (messageHistory.containsKey(packetId) || messageHistory.containsKey(messageId)) return@tryAndReport
+            messageHistory[packetId] = System.currentTimeMillis()
             inputStream.apply {
                 val isBroadcast = read() == 1
                 val endpoint = Endpoint.byId(read().toByte()) ?: return@apply
@@ -87,13 +91,14 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
                 }
                 if (messageBuilder.addPart(currentSlice, data)) {
                     messageBuilders.remove(messageId)
-                    receivedQueue.put(messageBuilder)
-                    networkHistory[messageId] = System.currentTimeMillis()
+                    processingQueue.put(messageBuilder)
+                    messageHistory[messageId] = System.currentTimeMillis()
                 }
             }
         }
     }
 
+    /**Sends outgoing [Message] from [outgoingQueue]. */
     private fun sendUDP() {
         val dataBuffer = ByteBuffer.allocate(configuration.packetSplitSize)
         while (true) tryAndReport {
@@ -147,23 +152,32 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
         }
     }
 
+    /** Processes received messages from [processingQueue] using [onMessageReceived] function.*/
     private fun processReceivedMessages() {
         while (true) {
-            val messageBuilder = receivedQueue.take()
+            val messageBuilder = processingQueue.take()
             onMessageReceived(messageBuilder.endpoint, messageBuilder.gluedData())
         }
     }
 
-    abstract fun onMessageReceived(endpoint: Endpoint, data: ByteArray)
+    /** Schedules message history cleanup service that runs at fixed rate. */
+    private fun startHistoryCleanup() {
+        Executors.newSingleThreadScheduledExecutor().scheduleAtFixedRate({
+            messageHistory.forEach { (messageHex, timestamp) ->
+                val difference = System.currentTimeMillis() - timestamp
+                val shouldBeRemoved = TimeUnit.MILLISECONDS.toMinutes(difference) >= configuration.historyMinuteClearance
+                if (shouldBeRemoved) messageHistory.remove(messageHex)
+            }
+        }, 0, configuration.historyCleaningFrequency, TimeUnit.MINUTES)
+    }
 
-    private class OutgoingQueuedMessage(
-        val endpoint: Endpoint,
-        val transmissionType: TransmissionType,
-        val messageUID: ByteArray,
-        val message: ByteArray,
-        val recipient: Node
-    )
+    /** Returns [Configuration.broadcastSpreadPercentage] number of nodes.  */
+    fun pickRandomNodes(amount: Int = 0): List<Node> {
+        val toTake = if (amount > 0) amount else 5 + (configuration.broadcastSpreadPercentage * Integer.max(totalKnownNodes, 1) / 100)
+        return getRandomNodes(toTake).filter { it != localNode }
+    }
 
+    /** Puts the action of adding the message to [outgoingQueue] in [queryStorage] for when the Node is found. Then the message is sent to the Node.*/
     fun send(endpoint: Endpoint, transmissionType: TransmissionType, message: Message, encodedMessage: ByteArray, vararg publicKeys: String) {
         val keys = publicKeys.takeIf { it.isNotEmpty() } ?: pickRandomNodes().map { it.publicKey }.toTypedArray()
         keys.forEach { key ->
@@ -176,21 +190,19 @@ abstract class Server(val configuration: Configuration) : Kademlia(configuration
         }
     }
 
-    /** Schedules message history cleanup service that runs at fixed rate. */
-    private fun startHistoryCleanup() {
-        Executors.newSingleThreadScheduledExecutor().scheduleAtFixedRate({
-            networkHistory.forEach { (messageHex, timestamp) ->
-                val difference = System.currentTimeMillis() - timestamp
-                val shouldBeRemoved = TimeUnit.MILLISECONDS.toMinutes(difference) >= configuration.historyMinuteClearance
-                if (shouldBeRemoved) networkHistory.remove(messageHex)
-            }
-        }, 0, configuration.historyCleaningFrequency, TimeUnit.MINUTES)
+    /** Picks [amount] of random closest nodes and sends them the message.*/
+    inline fun <reified T> send(endpoint: Endpoint, transmissionType: TransmissionType, data: T, amount: Int) {
+        val nodes = pickRandomNodes(amount)
+        send(endpoint, transmissionType, data, *nodes.map { it.publicKey }.toTypedArray())
     }
 
-    /** Returns [Configuration.broadcastSpreadPercentage] number of nodes.  */
-    fun pickRandomNodes(amount: Int = 0): List<Node> {
-        val toTake = if (amount > 0) amount else 5 + (configuration.broadcastSpreadPercentage * Integer.max(totalKnownNodes, 1) / 100)
-        return getRandomNodes(toTake).filter { it != localNode }
+    /** Computes the encoded message (Using ProtoBuf) and uses [send] to query for the Nodes using Kademlia protocol. */
+    inline fun <reified T> send(endpoint: Endpoint, transmissionType: TransmissionType, data: T, vararg publicKeys: String) {
+        val encodedBody = ProtoBuf.encodeToByteArray(data)
+        val signature = crypto.sign(encodedBody)
+        val message = Message(endpoint, crypto.publicKey, encodedBody, signature)
+        val encodedMessage = ProtoBuf.encodeToByteArray(message)
+        send(endpoint, transmissionType, message, encodedMessage, *publicKeys)
     }
 
 }
