--- src/test/kotlin/KademliaTest.kt
@@ -1,8 +1,0 @@
-/**
- * Created by mihael
- * on 02/12/2021 at 22:02
- * using IntelliJ IDEA
- */
-class KademliaTest {
-
-}
